- added sparx extra health in midnight when 117 category default enabled

+++ src/display_portal_verts.c
@@ -36,8 +36,8 @@ typedef struct
 } PolyF3;
 
 static CollisionHeader* cached_collision;
-static u32 cached_level = (u32)-1;
-static u32 cached_sublevel = (u32)-1;
+static u16 cached_level = -1;
+static u16 cached_sublevel = -1;
 static u16 portal_triangles[MAX_PORTAL_TRIANGLES];
 static u8 portal_triangle_count;
